% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_protein_name.R
\name{check_protein_name}
\alias{check_protein_name}
\title{Check if the protein name is valid}
\usage{
check_protein_name(
  protein_name,
 
    kierczak_et_al_2022_table_s2_xlsx_filename = get_local_kierczak_et_al_2022_table_s2_xlsx_filename(),
  url = get_kierczak_et_al_2022_table_s2_url()
)
}
\arguments{
\item{url}{the URL}
}
\value{
nothing
}
\description{
Check if the protein name is valid.
Will \link{stop} if not.
}
\examples{
check_protein_name(protein_name = "IL-6RA")
}
\author{
Richèl J.C. Bilderbeek
}
