% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fake_bianca.R
\name{create_fake_bianca}
\alias{create_fake_bianca}
\title{Fake the environment on Bianca as realistic as possible,
so the 'scripts_bianca' scripts can be run locally}
\usage{
create_fake_bianca()
}
\description{
Fake the environment on Bianca as realistic as possible,
so the 'scripts_bianca' scripts can be run locally
}
\author{
Richèl J.C. Bilderbeek
}
