% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_data_plots.R
\name{create_input_data_plots}
\alias{create_input_data_plots}
\title{Create plots about the input/GWAS data}
\usage{
create_input_data_plots(gcae_experiment_params)
}
\arguments{
\item{gcae_experiment_params}{parameters to run a full
\code{GCAE} experimenr (i.e. call \link[gcaer]{do_gcae_experiment}),
as created by \link[gcaer]{create_gcae_experiment_params} and checked
by \link[gcaer]{check_gcae_experiment_params}}
}
\value{
the filenames of the plots produced
}
\description{
Create plots about the input/GWAS data.
}
\author{
Richèl J.C. Bilderbeek
}
