% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_dataset.R
\name{create_random_dataset}
\alias{create_random_dataset}
\title{Create the \code{random} dataset, as described at
\url{https://github.com/AJResearchGroup/article_nsphs_ml_qt}}
\usage{
create_random_dataset(n_individuals, n_uniform, n_normal)
}
\arguments{
\item{n_individuals}{number of individuals to simulate}

\item{n_uniform}{number of traits with a random uniform
distribution to simulate}

\item{n_normal}{number of traits with a random normal
distribution to simulate}
}
\value{
an \code{assoc_qt_params} as
created by \link[plinkr]{create_demo_assoc_qt_params}
}
\description{
Create the \code{random} dataset, as described at
\url{https://github.com/AJResearchGroup/article_nsphs_ml_qt}
}
