% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_setting_1.R
\name{create_setting_1}
\alias{create_setting_1}
\title{Create setting 1 for GCAE}
\usage{
create_setting_1(base_input_filename = "setting_1")
}
\arguments{
\item{base_input_filename}{base filename of the files to be created}
}
\value{
a \link{list} with the following elements:
\itemize{
\item \code{bed_filename}: contains path to the \code{.bed} file
\item \code{bim_filename}: contains path to the \code{.bim} file
\item \code{fam_filename}: contains path to the \code{.fam} file
\item \code{phe_filename}: contains path to the \code{.phe} file
\item \code{labels_filename}: contains path to the file with the labels
}
}
\description{
\itemize{
\item Date: 2022-01-17
\item Goal: do the simplest useful simulation possible
\item Hypothesis: on the dimensionality reduction plots,
there will be three points.
With 'points' I really mean points, i.e. not clusters,
as there are exactly 3 genotypes
\item n = 1000
}
}
\details{
\tabular{lllll}{
   Genotype \tab Phenotype \tab Allele frequency \tab Population \tab Superpopulation \cr
   AA \tab 0 \tab 25\% \tab A \tab Global \cr
   AC \tab 50 \tab 50\% \tab B \tab Global \cr
   CC \tab 100 \tab 25\% \tab C \tab Global \cr
}
}
\examples{
# Create the files
filenames <- create_setting_1()

# Clean up
file.remove(as.character(unlist(filenames)))
}
\author{
Richèl J.C. Bilderbeek
}
