% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(gcae_experiment_params, verbose)
}
\arguments{
\item{gcae_experiment_params}{parameters to run a full
\code{GCAE} experimenr (i.e. call \link[gcaer]{do_gcae_experiment}),
as created by \link[gcaer]{create_gcae_experiment_params} and checked
by \link[gcaer]{check_gcae_experiment_params}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
