% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dna_sequences.R
\name{get_dna_sequences}
\alias{get_dna_sequences}
\title{Get DNA sequences}
\usage{
get_dna_sequences(n_people = 10)
}
\arguments{
\item{n_people}{the number of people}
}
\description{
Get DNA sequences
}
