% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_sequences.R
\name{get_genome_sequences}
\alias{get_genome_sequences}
\title{Get the genome sequences}
\usage{
get_genome_sequences(n_people = 10, sequence_length = 20)
}
\arguments{
\item{n_people}{the number of people}

\item{sequence_length}{the DNA sequence length}
}
\description{
Get the genome sequences
}
