% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genotype_map.R
\name{get_genotype_map}
\alias{get_genotype_map}
\title{Get the genotype map}
\usage{
get_genotype_map(n_people = 10, n_snvs = 5)
}
\arguments{
\item{n_people}{the number of people}

\item{n_snvs}{the number of single nucleotide variants}
}
\description{
Get the genotype map
}
