% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kierczak_et_al_2022_table_s2_url.R
\name{get_kierczak_et_al_2022_table_s2_url}
\alias{get_kierczak_et_al_2022_table_s2_url}
\title{Get the URL of Kierczak et al., 2002, for supplementary table 2}
\usage{
get_kierczak_et_al_2022_table_s2_url()
}
\value{
the URL
}
\description{
Get the URL of Kierczak et al., 2002, for supplementary table 2
}
\examples{
get_kierczak_et_al_2022_table_s2_url()
}
\author{
Richèl J.C. Bilderbeek
}
