% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kierczak_et_al_2022_table_s4_csv.R
\name{get_kierczak_et_al_2022_table_s4_csv}
\alias{get_kierczak_et_al_2022_table_s4_csv}
\title{Get Supplementary Table 4 of Kierczak et al., 2022
that was supplied as a CSV}
\usage{
get_kierczak_et_al_2022_table_s4_csv()
}
\description{
Get Supplementary Table 4 of Kierczak et al., 2022
that was supplied as a CSV
}
\note{
The article is in pre-print as of 2021. As I predict the publication
will be in 2022, I will use that in the function name.
Full citation (as now in Google Scholar): Kierczak, Marcin, et al.
"The contribution of rare whole genome sequencing variants to plasma protein
levels and to the missing heritability." (2021).
\url{10.21203/rs.3.rs-625433/v1}
}
