% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kierczak_et_al_2022_table_s4_xlsx.R
\name{get_kierczak_et_al_2022_table_s4_xlsx}
\alias{get_kierczak_et_al_2022_table_s4_xlsx}
\title{Get Supplementary Table 1 of Kierczak et al., 2022}
\usage{
get_kierczak_et_al_2022_table_s4_xlsx(
 
    kierczak_et_al_2022_table_s4_xlsx_filename = file.path(rappdirs::app_dir("nsphsmlqt")$data(),
    "TableS4.V3reduced.xlsx"),
 
    url = "https://assets.researchsquare.com/files/rs-625433/v1/3c21e68a0d1b15f652295e4b.xlsx"
)
}
\arguments{
\item{kierczak_et_al_2022_table_s4_xlsx_filename}{name of the file to
save the Excel sheet to}

\item{url}{the URL to download the Excel sheet from}
}
\description{
Get Supplementary Table 1 of Kierczak et al., 2022
}
\note{
The article is in pre-print as of 2021. As I predict the publication
will be in 2022, I will use that in the function name.
Pre-print of article is at URL \url{10.21203/rs.3.rs-625433/v1}
}
