% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_local_kierczak_et_al_2022_table_s2_xlsx_filename.R
\name{get_local_kierczak_et_al_2022_table_s2_xlsx_filename}
\alias{get_local_kierczak_et_al_2022_table_s2_xlsx_filename}
\title{Get the local path where the file downloaded from
\link{get_kierczak_et_al_2022_table_s2_url} will be stored.}
\usage{
get_local_kierczak_et_al_2022_table_s2_xlsx_filename()
}
\value{
the filename
}
\description{
Get the local path where the file downloaded from
\link{get_kierczak_et_al_2022_table_s2_url} will be stored.
}
\examples{
get_local_kierczak_et_al_2022_table_s2_xlsx_filename()
}
\author{
Richèl J.C. Bilderbeek
}
