% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_primary_signal.R
\name{get_primary_signal}
\alias{get_primary_signal}
\title{Get the primary cis association for a protein}
\usage{
get_primary_signal(
  protein_name,
 
    kierczak_et_al_2022_table_s2_xlsx_filename = get_local_kierczak_et_al_2022_table_s2_xlsx_filename(),
  url = get_kierczak_et_al_2022_table_s2_url()
)
}
\arguments{
\item{url}{the URL}
}
\value{
a \link[tibble]{tibble}
}
\description{
Get the primary cis association for a protein
}
\examples{
get_primary_signal(protein_name = "IL-6RA")
}
\author{
Richèl J.C. Bilderbeek
}
