% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_genome_sequence.R
\name{get_reference_genome_sequence}
\alias{get_reference_genome_sequence}
\title{Get the reference genome sequence}
\usage{
get_reference_genome_sequence(n_nucleotides = 20)
}
\arguments{
\item{n_nucleotides}{the length of the sequences, in the
number of nucleotides}
}
\description{
Get the reference genome sequence
}
